Move WebsocketRaidsHandler to actor package
